requirements.txt
torch
wandb
numpy
pandas
hydra-core
timm

# dvc
dvc
dvc_gdrive
hydra-core
dvc_gs

python-dotenv
fastapi
uvicorn
pydantic
pillow

transformers
python-multipart
scikit-learn
```